Modified core data to use the JCPhoneNumberDataSource protocol on recent events. Did away with "extension" in favor of a single value of "number" to be consistent throughout the restive the code base.

+++ JiveOne/JCSipManager.m
@@ -159,7 +159,7 @@ NSString *const kSipHandlerRegisteredSelectorKey = @"registered";
     int errorCode = [_mPortSIPSDK initialize:TRANSPORT_UDP
                                     loglevel:LOG_LEVEL
                                      logPath:nil
-                                     maxLine:_numberOfLines
+                                     maxLine:(int)_numberOfLines
                                        agent:kSipHandlerServerAgentname
                             audioDeviceLayer:IS_SIMULATOR
                             videoDeviceLayer:IS_SIMULATOR];
